requirements.txt
pandas
supabase
streamlit-authenticator
openai
langchain
langsmith
langchain_community
langchain_openai
# langchain-chroma==0.1.2
chromadb
google-colab
pysqlite-binary 